+++ freemarker-generator-maven-plugin/src/main/java/org/apache/freemarker/generator/maven/GeneratingFileVisitor.java
@@ -16,7 +16,6 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
 package org.apache.freemarker.generator.maven;
 
 import freemarker.template.Configuration;
@@ -30,7 +29,7 @@ import java.util.Map;
 
 /**
  * FileVisitor designed to process json data files. The json file parsed into
- * a map and given to FreeMarker to
+ * a map and given to FreeMarker to.
  */
 public class GeneratingFileVisitor extends SimpleFileVisitor<Path> {
 
@@ -52,8 +51,8 @@ public class GeneratingFileVisitor extends SimpleFileVisitor<Path> {
     /**
      * Factory method that calls constructor, added to facilitate testing with jmockit.
      *
-     * @param config FreeMarker configuration
-     * @param session Maven session
+     * @param config             FreeMarker configuration
+     * @param session            Maven session
      * @param extensionToBuilder builder extension
      * @return GeneratingFileVisitor instance
      */
@@ -68,8 +67,8 @@ public class GeneratingFileVisitor extends SimpleFileVisitor<Path> {
                     .addGeneratorLocation(path)
                     .addPomLastModifiedTimestamp(pomLastModifiedTimestamp);
             final String fileName = path.getFileName().toString();
-            final String extenstion = fileName.substring(fileName.lastIndexOf('.'));
-            final OutputGeneratorPropertiesProvider pathProcessor = extensionToBuilder.get(extenstion);
+            final String extension = fileName.substring(fileName.lastIndexOf('.'));
+            final OutputGeneratorPropertiesProvider pathProcessor = extensionToBuilder.get(extension);
             if (pathProcessor == null) {
                 throw new RuntimeException("Unknown file extension: " + path);
             }