+++ freemarker-generator-maven-plugin/src/test/java/org/apache/freemarker/generator/maven/FreeMarkerMojoTest.java
@@ -16,346 +16,166 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
 package org.apache.freemarker.generator.maven;
 
-import freemarker.cache.FileTemplateLoader;
-import freemarker.cache.TemplateLoader;
-import freemarker.template.Configuration;
-import mockit.Deencapsulation;
-import mockit.Expectations;
-import mockit.Mocked;
-import mockit.Verifications;
+import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
+import org.apache.commons.lang3.reflect.FieldUtils;
+import org.apache.freemarker.generator.base.util.OperatingSystem;
 import org.apache.maven.execution.MavenSession;
 import org.apache.maven.plugin.MojoExecution;
 import org.apache.maven.plugin.MojoExecutionException;
-import org.apache.maven.plugin.MojoFailureException;
 import org.apache.maven.project.MavenProject;
 import org.junit.Assert;
-import org.testng.annotations.BeforeClass;
-import org.testng.annotations.Test;
+import org.junit.BeforeClass;
+import org.junit.Test;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.FileVisitor;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.Comparator;
-import java.util.Map;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Properties;
 
 import static org.assertj.core.api.Assertions.*;
+import static org.mockito.Mockito.*;
 
 public class FreeMarkerMojoTest extends Assert {
 
     private static final File TEST_OUTPUT_DIR = new File("target/test-output/freemarker-mojo");
-    private static final String FREEMARKER_VERSION = "2.3.29";
+    private static final String FREEMARKER_VERSION = "2.3.30";
 
     @BeforeClass
     public static void beforeClass() throws IOException {
-        // Clean output dir before each run.
-        if (TEST_OUTPUT_DIR.exists()) {
-            // Recursively delete output from previous run.
-            Files.walk(TEST_OUTPUT_DIR.toPath())
-                    .sorted(Comparator.reverseOrder())
-                    .map(Path::toFile)
-                    .forEach(File::delete);
-        }
+        UnitTestHelper.deleteTestOutputDir(TEST_OUTPUT_DIR);
     }
 
     @Test
-    public void executeTest(
-            @Mocked MavenSession session,
-            @Mocked MavenProject project,
-            @Mocked MojoExecution mojoExecution,
-            @Mocked GeneratingFileVisitor generatingFileVisitor,
-            @Mocked Files files
-    ) throws MojoExecutionException, MojoFailureException, IOException {
+    public void execute_generateSourceTest() throws MojoExecutionException, IOException, IllegalAccessException {
 
-        new Expectations(mojoExecution, generatingFileVisitor) {{
-            mojoExecution.getLifecyclePhase();
-            result = "generate-sources";
-            session.getCurrentProject();
-            result = project;
-        }};
+        final File mockFile = mock(File.class);
+        final MavenSession session = mock(MavenSession.class);
+        final MavenProject project = mock(MavenProject.class);
+        final MojoExecution mojoExecution = mock(MojoExecution.class);
 
-        FreeMarkerMojo mojo = new FreeMarkerMojo();
+        final List<MavenProject> projects = new ArrayList<>();
+        projects.add(project);
+
+        when(session.getAllProjects()).thenReturn(projects);
+        when(session.getCurrentProject()).thenReturn(project);
+        when(session.getCurrentProject().getProperties()).thenReturn(new Properties());
+        when(project.getFile()).thenReturn(mockFile);
+        when(mockFile.lastModified()).thenReturn(10L);
+        when(mojoExecution.getLifecyclePhase()).thenReturn("generate-sources");
+
+        final FreeMarkerMojo mojo = new FreeMarkerMojo();
+        final File sourceDirectory = new File("src/test/data/freemarker-mojo");
+        final File testCaseOutputDir = new File(TEST_OUTPUT_DIR, "executeTest");
+
+        FieldUtils.writeField(mojo, "freeMarkerVersion", "", true);
+        FieldUtils.writeField(mojo, "freeMarkerVersion", FREEMARKER_VERSION, true);
+        FieldUtils.writeField(mojo, "sourceDirectory", sourceDirectory, true);
+        FieldUtils.writeField(mojo, "templateDirectory", new File(testCaseOutputDir, "template"), true);
+        FieldUtils.writeField(mojo, "generatorDirectory", new File(testCaseOutputDir, "data"), true);
+        FieldUtils.writeField(mojo, "outputDirectory", new File(testCaseOutputDir, "generated-files"), true);
+        FieldUtils.writeField(mojo, "mojo", mojoExecution, true);
+        FieldUtils.writeField(mojo, "session", session, true);
+
+        FileUtils.forceMkdir(new File(testCaseOutputDir, "data"));
+        FileUtils.forceMkdir(new File(testCaseOutputDir, "template"));
+
+        mojo.execute();
+
+        verify(project).addCompileSourceRoot(fixSeparators("target/test-output/freemarker-mojo/executeTest/generated-files"));
+        verify(project, times(0)).addTestCompileSourceRoot(anyString());
+    }
+
+    @Test
+    public void execute_generateTestSourceTest() throws MojoExecutionException, IOException, IllegalAccessException {
+
+        final File mockFile = mock(File.class);
+        final MavenSession session = mock(MavenSession.class);
+        final MavenProject project = mock(MavenProject.class);
+        final MojoExecution mojoExecution = mock(MojoExecution.class);
+
+        final List<MavenProject> projects = new ArrayList<>();
+        projects.add(project);
+
+        when(session.getAllProjects()).thenReturn(projects);
+        when(session.getCurrentProject()).thenReturn(project);
+        when(session.getCurrentProject().getProperties()).thenReturn(new Properties());
+        when(project.getFile()).thenReturn(mockFile);
+        when(mockFile.lastModified()).thenReturn(10L);
+        when(mojoExecution.getLifecyclePhase()).thenReturn("generate-test-sources");
+
+        final FreeMarkerMojo mojo = new FreeMarkerMojo();
+        final File sourceDirectory = new File("src/test/data/freemarker-mojo");
+        final File testCaseOutputDir = new File(TEST_OUTPUT_DIR, "generateTestSourceTest");
+
+        FieldUtils.writeField(mojo, "freeMarkerVersion", "", true);
+        FieldUtils.writeField(mojo, "freeMarkerVersion", FREEMARKER_VERSION, true);
+        FieldUtils.writeField(mojo, "sourceDirectory", sourceDirectory, true);
+        FieldUtils.writeField(mojo, "templateDirectory", new File(testCaseOutputDir, "template"), true);
+        FieldUtils.writeField(mojo, "generatorDirectory", new File(testCaseOutputDir, "data"), true);
+        FieldUtils.writeField(mojo, "outputDirectory", new File(testCaseOutputDir, "generated-files"), true);
+        FieldUtils.writeField(mojo, "mojo", mojoExecution, true);
+        FieldUtils.writeField(mojo, "session", session, true);
+
+        FileUtils.forceMkdir(new File(testCaseOutputDir, "data"));
+        FileUtils.forceMkdir(new File(testCaseOutputDir, "template"));
+
+        mojo.execute();
+
+        verify(project, times(0)).addCompileSourceRoot(anyString());
+        verify(project).addTestCompileSourceRoot(fixSeparators("target/test-output/freemarker-mojo/generateTestSourceTest/generated-files"));
+    }
+
+    @Test
+    public void execute_checkPluginParametersTest()
+            throws MojoExecutionException, IOException, IllegalAccessException {
+
+        UnitTestHelper.deleteTestOutputDir(TEST_OUTPUT_DIR);
+
+        final MavenSession session = mock(MavenSession.class);
+        final MojoExecution mojoExecution = mock(MojoExecution.class);
+        final FreeMarkerMojo mojo = new FreeMarkerMojo();
 
         // Validate freeMarkerVersion is required.
-        assertThatExceptionOfType(MojoExecutionException.class).isThrownBy(() -> {
-            mojo.execute();
-        }).withMessage("freeMarkerVersion is required");
 
-        Deencapsulation.setField(mojo, "freeMarkerVersion", "");
-        assertThatExceptionOfType(MojoExecutionException.class).isThrownBy(() -> {
-            mojo.execute();
-        }).withMessage("freeMarkerVersion is required");
+        FieldUtils.writeField(mojo, "freeMarkerVersion", null, true);
+        assertThatExceptionOfType(MojoExecutionException.class)
+                .isThrownBy(mojo::execute)
+                .withMessage("freeMarkerVersion is required");
+
+        FieldUtils.writeField(mojo, "freeMarkerVersion", "", true);
+        assertThatExceptionOfType(MojoExecutionException.class)
+                .isThrownBy(mojo::execute)
+                .withMessage("freeMarkerVersion is required");
 
         final File testCaseOutputDir = new File(TEST_OUTPUT_DIR, "executeTest");
-        Deencapsulation.setField(mojo, "freeMarkerVersion", FREEMARKER_VERSION);
-        Deencapsulation.setField(mojo, "sourceDirectory", testCaseOutputDir);
-        Deencapsulation.setField(mojo, "templateDirectory", new File(testCaseOutputDir, "template"));
-        Deencapsulation.setField(mojo, "generatorDirectory", new File(testCaseOutputDir, "data"));
-        Deencapsulation.setField(mojo, "outputDirectory", new File(testCaseOutputDir, "generated-files"));
-        Deencapsulation.setField(mojo, "mojo", mojoExecution);
-        Deencapsulation.setField(mojo, "session", session);
+        FieldUtils.writeField(mojo, "freeMarkerVersion", FREEMARKER_VERSION, true);
+        FieldUtils.writeField(mojo, "sourceDirectory", testCaseOutputDir, true);
+        FieldUtils.writeField(mojo, "templateDirectory", new File(testCaseOutputDir, "template"), true);
+        FieldUtils.writeField(mojo, "generatorDirectory", new File(testCaseOutputDir, "data"), true);
+        FieldUtils.writeField(mojo, "outputDirectory", new File(testCaseOutputDir, "generated-files"), true);
+        FieldUtils.writeField(mojo, "mojo", mojoExecution, true);
+        FieldUtils.writeField(mojo, "session", session, true);
 
         // Validate source directory.
-        assertThatExceptionOfType(MojoExecutionException.class).isThrownBy(() -> {
-            mojo.execute();
-        })
-                .withMessageStartingWith("Required directory does not exist");
 
-        new File(testCaseOutputDir, "data").mkdirs();
-        assertThatExceptionOfType(MojoExecutionException.class).isThrownBy(() -> {
-            mojo.execute();
-        })
-                .withMessageStartingWith("Required directory does not exist");
-        new File(testCaseOutputDir, "template").mkdirs();
+        assertThatExceptionOfType(MojoExecutionException.class)
+                .isThrownBy(mojo::execute)
+                .withMessageStartingWith("Required generator directory does not exist");
+
+        FileUtils.forceMkdir(new File(testCaseOutputDir, "data"));
+        assertThatExceptionOfType(MojoExecutionException.class)
+                .isThrownBy(mojo::execute)
+                .withMessageStartingWith("Required template directory does not exist");
+
+        FileUtils.forceMkdir(new File(testCaseOutputDir, "template"));
 
         // Validate minimum configuration.
-        mojo.execute();
-
-        new Verifications() {{
-            project.addCompileSourceRoot(fixSeparators("target/test-output/freemarker-mojo/executeTest/generated-files"));
-            times = 1;
-
-            Configuration config;
-            MavenSession capturedSession;
-            Map<String, OutputGeneratorPropertiesProvider> builders;
-
-            GeneratingFileVisitor.create(
-                    config = withCapture(),
-                    capturedSession = withCapture(),
-                    builders = withCapture());
-            times = 1;
-
-            assertEquals("UTF-8", config.getDefaultEncoding());
-            assertEquals(session, capturedSession);
-            TemplateLoader loader = config.getTemplateLoader();
-            assertTrue(loader instanceof FileTemplateLoader);
-
-            Path path;
-            FileVisitor<Path> fileVisitor;
-
-            Files.walkFileTree(path = withCapture(), fileVisitor = withCapture());
-            times = 1;
-
-            assertEquals(new File(testCaseOutputDir, "data").toPath(), path);
-            assertTrue(fileVisitor instanceof GeneratingFileVisitor);
-        }};
-    }
-
-    @Test
-    public void execute_generateTestSourceTest(
-            @Mocked MavenSession session,
-            @Mocked MavenProject project,
-            @Mocked MojoExecution mojoExecution,
-            @Mocked GeneratingFileVisitor generatingFileVisitor,
-            @Mocked Files files
-    ) throws MojoExecutionException, MojoFailureException, IOException {
-
-        new Expectations(mojoExecution, generatingFileVisitor) {{
-            mojoExecution.getLifecyclePhase();
-            result = "generate-test-sources";
-            session.getCurrentProject();
-            result = project;
-        }};
-
-        FreeMarkerMojo mojo = new FreeMarkerMojo();
-
-        File testCaseOutputDir = new File(TEST_OUTPUT_DIR, "generateTestSourceTest");
-        Deencapsulation.setField(mojo, "freeMarkerVersion", FREEMARKER_VERSION);
-        Deencapsulation.setField(mojo, "sourceDirectory", testCaseOutputDir);
-        Deencapsulation.setField(mojo, "templateDirectory", new File(testCaseOutputDir, "template"));
-        Deencapsulation.setField(mojo, "generatorDirectory", new File(testCaseOutputDir, "data"));
-        Deencapsulation.setField(mojo, "outputDirectory", new File(testCaseOutputDir, "generated-files"));
-        Deencapsulation.setField(mojo, "mojo", mojoExecution);
-        Deencapsulation.setField(mojo, "session", session);
-
-        new File(testCaseOutputDir, "data").mkdirs();
-        new File(testCaseOutputDir, "template").mkdirs();
 
         mojo.execute();
-
-        new Verifications() {{
-            project.addTestCompileSourceRoot(fixSeparators("target/test-output/freemarker-mojo/generateTestSourceTest/generated-files"));
-            times = 1;
-        }};
-    }
-
-    @SuppressWarnings({ "rawtypes", "unchecked" })
-    @Test
-    public void execute_walkFileTreeExceptionTest(
-            @Mocked MavenSession session,
-            @Mocked MavenProject project,
-            @Mocked MojoExecution mojoExecution,
-            @Mocked GeneratingFileVisitor generatingFileVisitor,
-            @Mocked Files files
-    ) throws MojoExecutionException, MojoFailureException, IOException {
-
-        new Expectations(mojoExecution, generatingFileVisitor) {{
-            mojoExecution.getLifecyclePhase();
-            result = "generate-test-sources";
-            session.getCurrentProject();
-            result = project;
-            Files.walkFileTree((Path) any, (FileVisitor) any);
-            result = new RuntimeException("test exception");
-        }};
-
-        FreeMarkerMojo mojo = new FreeMarkerMojo();
-
-        File testCaseOutputDir = new File(TEST_OUTPUT_DIR, "generateTestSourceTest");
-        Deencapsulation.setField(mojo, "freeMarkerVersion", FREEMARKER_VERSION);
-        Deencapsulation.setField(mojo, "sourceDirectory", testCaseOutputDir);
-        Deencapsulation.setField(mojo, "templateDirectory", new File(testCaseOutputDir, "template"));
-        Deencapsulation.setField(mojo, "generatorDirectory", new File(testCaseOutputDir, "data"));
-        Deencapsulation.setField(mojo, "outputDirectory", new File(testCaseOutputDir, "generated-files"));
-        Deencapsulation.setField(mojo, "mojo", mojoExecution);
-        Deencapsulation.setField(mojo, "session", session);
-
-        new File(testCaseOutputDir, "data").mkdirs();
-        new File(testCaseOutputDir, "template").mkdirs();
-
-        assertThatExceptionOfType(MojoExecutionException.class).isThrownBy(() -> {
-            mojo.execute();
-        })
-                .withMessageStartingWith("Failed to process files in generator dir");
-    }
-
-    @Test
-    public void execute_setTemplateLoaderExceptionTest(
-            @Mocked MavenSession session,
-            @Mocked MavenProject project,
-            @Mocked MojoExecution mojoExecution,
-            @Mocked FactoryUtil factoryUtil,
-            @Mocked Configuration config) {
-
-        new Expectations(config, FactoryUtil.class) {{
-            FactoryUtil.createConfiguration(FREEMARKER_VERSION);
-            result = config;
-            config.setTemplateLoader((TemplateLoader) any);
-            result = new RuntimeException("test exception");
-        }};
-
-        FreeMarkerMojo mojo = new FreeMarkerMojo();
-
-        File testCaseOutputDir = new File(TEST_OUTPUT_DIR, "setTemplateLoaderException");
-
-        Deencapsulation.setField(mojo, "freeMarkerVersion", FREEMARKER_VERSION);
-        Deencapsulation.setField(mojo, "sourceDirectory", testCaseOutputDir);
-        Deencapsulation.setField(mojo, "templateDirectory", new File(testCaseOutputDir, "template"));
-        Deencapsulation.setField(mojo, "generatorDirectory", new File(testCaseOutputDir, "data"));
-        Deencapsulation.setField(mojo, "outputDirectory", new File(testCaseOutputDir, "generated-files"));
-        Deencapsulation.setField(mojo, "mojo", mojoExecution);
-        Deencapsulation.setField(mojo, "session", session);
-
-        new File(testCaseOutputDir, "data").mkdirs();
-        new File(testCaseOutputDir, "template").mkdirs();
-
-        assertThatExceptionOfType(MojoExecutionException.class).isThrownBy(() -> {
-            mojo.execute();
-        })
-                .withMessageStartingWith("Could not establish file template loader for directory");
-    }
-
-    @Test
-    public void execute_loadFreemarkerPropertiesTest(
-            @Mocked MavenSession session,
-            @Mocked MavenProject project,
-            @Mocked MojoExecution mojoExecution,
-            @Mocked Configuration config) throws Exception {
-
-        FreeMarkerMojo mojo = new FreeMarkerMojo();
-
-        File sourceDirectory = new File("src/test/data/freemarker-mojo");
-        File testCaseOutputDir = new File(TEST_OUTPUT_DIR, "loadFreemarkerProperties");
-
-        Deencapsulation.setField(mojo, "freeMarkerVersion", FREEMARKER_VERSION);
-        Deencapsulation.setField(mojo, "sourceDirectory", sourceDirectory);
-        Deencapsulation.setField(mojo, "templateDirectory", new File(sourceDirectory, "template"));
-        Deencapsulation.setField(mojo, "generatorDirectory", new File(sourceDirectory, "data"));
-        Deencapsulation.setField(mojo, "outputDirectory", new File(testCaseOutputDir, "generated-files"));
-        Deencapsulation.setField(mojo, "mojo", mojoExecution);
-        Deencapsulation.setField(mojo, "session", session);
-
-        mojo.execute();
-
-        new Verifications() {{
-            Properties properties;
-
-            config.setSettings(properties = withCapture());
-            times = 1;
-
-            assertEquals("T,F", properties.getProperty("boolean_format"));
-        }};
-    }
-
-    @Test
-    public void execute_loadFreemarkerPropertiesExceptionTest(
-            @Mocked MavenSession session,
-            @Mocked MavenProject project,
-            @Mocked MojoExecution mojoExecution,
-            @Mocked FactoryUtil factoryUtil,
-            @Mocked Configuration config) throws Exception {
-
-        new Expectations(FactoryUtil.class) {{
-            FactoryUtil.createFileInputStream((File) any);
-            result = new RuntimeException("test exception");
-        }};
-
-        FreeMarkerMojo mojo = new FreeMarkerMojo();
-
-        File sourceDirectory = new File("src/test/data/freemarker-mojo");
-        File testCaseOutputDir = new File(TEST_OUTPUT_DIR, "loadFreemarkerPropertiesExceptionTest");
-
-        Deencapsulation.setField(mojo, "freeMarkerVersion", FREEMARKER_VERSION);
-        Deencapsulation.setField(mojo, "sourceDirectory", sourceDirectory);
-        Deencapsulation.setField(mojo, "templateDirectory", new File(sourceDirectory, "template"));
-        Deencapsulation.setField(mojo, "generatorDirectory", new File(sourceDirectory, "data"));
-        Deencapsulation.setField(mojo, "outputDirectory", new File(testCaseOutputDir, "generated-files"));
-        Deencapsulation.setField(mojo, "mojo", mojoExecution);
-        Deencapsulation.setField(mojo, "session", session);
-
-        System.out.println("==== before mojo execute");
-        try {
-            assertThatExceptionOfType(MojoExecutionException.class).isThrownBy(() -> {
-                mojo.execute();
-            }).withMessage(fixSeparators("Failed to load src/test/data/freemarker-mojo/freemarker.properties"));
-        } catch (Throwable t) {
-            t.printStackTrace();
-        }
-    }
-
-    @Test
-    public void execute_setSettingsExceptionTest(
-            @Mocked MavenSession session,
-            @Mocked MavenProject project,
-            @Mocked MojoExecution mojoExecution,
-            @Mocked Configuration config) throws Exception {
-
-        new Expectations() {{
-            config.setSettings((Properties) any);
-            result = new RuntimeException("test exception");
-        }};
-
-        FreeMarkerMojo mojo = new FreeMarkerMojo();
-
-        File sourceDirectory = new File("src/test/data/freemarker-mojo");
-        File testCaseOutputDir = new File(TEST_OUTPUT_DIR, "loadFreemarkerProperties");
-
-        Deencapsulation.setField(mojo, "freeMarkerVersion", FREEMARKER_VERSION);
-        Deencapsulation.setField(mojo, "sourceDirectory", sourceDirectory);
-        Deencapsulation.setField(mojo, "templateDirectory", new File(sourceDirectory, "template"));
-        Deencapsulation.setField(mojo, "generatorDirectory", new File(sourceDirectory, "data"));
-        Deencapsulation.setField(mojo, "outputDirectory", new File(testCaseOutputDir, "generated-files"));
-        Deencapsulation.setField(mojo, "mojo", mojoExecution);
-        Deencapsulation.setField(mojo, "session", session);
-
-        assertThatExceptionOfType(MojoExecutionException.class).isThrownBy(() -> {
-            mojo.execute();
-        }).withMessage(fixSeparators("Invalid setting(s) in src/test/data/freemarker-mojo/freemarker.properties"));
     }
 
     private static String fixSeparators(String str) {